switch to KangarooTwelve proof of work

+++ src/cryptonote_basic/miner.cpp
@@ -131,7 +131,7 @@ namespace cryptonote
     m_diffic = di;
     m_height = height;
     ++m_template_no;
-    m_starter_nonce = crypto::rand<uint32_t>();
+    m_starter_nonce = crypto::rand<uint64_t>();
     return true;
   }
   //-----------------------------------------------------------------------------------------------------
@@ -294,7 +294,7 @@ namespace cryptonote
   {
     m_mine_address = adr;
     m_threads_total = static_cast<uint32_t>(threads_count);
-    m_starter_nonce = crypto::rand<uint32_t>();
+    m_starter_nonce = crypto::rand<uint64_t>();
     CRITICAL_REGION_LOCAL(m_threads_lock);
     if(is_mining())
     {
@@ -383,7 +383,7 @@ namespace cryptonote
   //-----------------------------------------------------------------------------------------------------
   bool miner::find_nonce_for_given_block(block& bl, const difficulty_type& diffic, uint64_t height)
   {
-    for(; bl.nonce != std::numeric_limits<uint32_t>::max(); bl.nonce++)
+    for(; bl.nonce != std::numeric_limits<uint64_t>::max(); bl.nonce++)
     {
       crypto::hash h;
       get_block_longhash(bl, h, height);
@@ -437,7 +437,7 @@ namespace cryptonote
     uint32_t th_local_index = boost::interprocess::ipcdetail::atomic_inc32(&m_thread_index);
     MLOG_SET_THREAD_NAME(std::string("[miner ") + std::to_string(th_local_index) + "]");
     MGINFO("Miner thread was started ["<< th_local_index << "]");
-    uint32_t nonce = m_starter_nonce + th_local_index;
+    uint64_t nonce = m_starter_nonce + th_local_index;
     uint64_t height = 0;
     difficulty_type local_diff = 0;
     uint32_t local_template_ver = 0;
